YAML to Mux: move all related classes to plugin
These classes and methods have knowledge about the YAML files, and the
YAML to Mux plugin is the only user of those.  this plugin.  So it
makes sense to move it to the plugin module.
